Edit code to check logged in user before delete comment (more secure) and change block comment
Change block comment to show error when user want to post more than one comment instead of block comment element.

+++ tinderforeduapp/views.py
@@ -50,7 +50,7 @@ def signup(request):
                                                   age=user.profile.age, firstname=user.profile.first_name,
                                                   lastname=user.profile.last_name, gender=user.profile.gender)
             # Set user's profile picture to default and link PictureContainer object to user's object
-            PictureContainer.objects.create(user=create_user, images='image_holder.png', is_profile_pic=True)
+            PictureContainer.objects.create(user=create_user, images='default_profile_image.png', is_profile_pic=True)
             # Save all objects
             create_user.save()
             user.save()
@@ -80,8 +80,10 @@ def delete_comment(request, user_id):
         comment_id = request.POST['comment_id']
         # find object of class Comment by id
         comment_object = Comment.objects.get(id=comment_id)
-        # delete it
-        comment_object.delete()
+        # check if selected object is belong to logged in user
+        if comment_object.name == request.user.username:
+            # delete it
+            comment_object.delete()
     return redirect('{}#comment_section_start'.format(reverse('tinder:matched_profile', args=(user_id,))))
 
 
@@ -490,15 +492,17 @@ def matched_profile(request, user_id):
     selected_user_profile_picture = selected_user_comment_object.image.filter(is_profile_pic=True)
     # get only active comment
     comments = selected_user_comment_object.comments.filter(active=True)
-    # send flag to tell this login user had comment to this selected user
-    block_comment = False
-    for comment in comments:
-        if request.user.username in comment.name:
-            block_comment = True
     additional_pic = selected_user_comment_object.image.filter(is_profile_pic=False)
     new_comment = None
     # if logged in user comment on selected user
     if request.method == 'POST':
+        # check if logged in user try to comment on commented profile
+        for comment in comments:
+            # if found comment by logged in user on selected user's profile
+            if request.user.username in comment.name:
+                # alert logged in user
+                return render(request, 'tinder/no_comment.html',
+                              {'user_id': user_id, 'url': get_current_site(request).domain})
         comment_form = CommentForm(data=request.POST)
         # if form valid
         if comment_form.is_valid():
@@ -532,7 +536,6 @@ def matched_profile(request, user_id):
                   {'pic': selected_user_profile_picture, 'additional_pic': additional_pic,
                    'name': UserInfo.objects.get(name=request.user.username),
                    'username': request.user.username,
-                   'enable_comment': block_comment,
                    'profile': UserInfo.objects.get(id=user_id), 'post': selected_user_comment_object,
                    'comments': comments, 'new_comment': new_comment, 'comment_form': comment_form})
